+++ src/librustc/query/mod.rs
@@ -630,6 +630,10 @@ rustc_queries! {
                                         -> &'tcx [(CrateNum, LinkagePreference)] {
             desc { "dylib dependency formats of crate" }
         }
+
+        query dependency_formats(_: CrateNum) -> Lrc<crate::middle::dependency_format::Dependencies> {
+            desc { "get the linkage format of all dependencies" }
+        }
     }
 
     Codegen {
